requirements.txt
accelerate>=1.13.0
numpy==2.2.6; python_version < "3.14"
numpy>=2.3.0; python_version >= "3.14"
torch>=2.8.0
safetensors>=0.7.0
transformers>=5.4.0
threadpoolctl>=3.6.0
packaging>=24.2
device-smi>=0.5.5
protobuf>=7.34.0
pillow>=11.3.0
pypcre>=0.3.2
tokenicer>=0.0.13
logbar>=0.4.3
jinja2>=3.1.0
ninja>=1.13.0
maturin>=1.9.4
datasets>=3.6.0
pyarrow>=21.0
dill>=0.3.8
torchao>=0.16.0
defuser>=0.0.21
